Extract data classes to separate files
