+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/receipts/ReceiptCreator.kt
@@ -135,44 +135,4 @@ class ReceiptCreator {
         const val SUMMARY_SECTION_TITLE = "HC: Summary"
         const val AID = "HC: AID"
     }
-
-    enum class CardBrand(val iconName: String, val base64Data: Lazy<String>) {
-        /* ktlint-disable max-line-length */
-        VISA(
-            iconName = "visa",
-            base64Data = lazyOf(
-                "PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiIHN0YW5kYWxvbmU9Im5vIj8+DQo8c3ZnIHZpZXdCb3g9IjAgMCA3NTAgNDcxIiB2ZXJzaW9uPSIxLjEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHhtbG5zOnNrZXRjaD0iaHR0cDovL3d3dy5ib2hlbWlhbmNvZGluZy5jb20vc2tldGNoL25zIiBwcmVzZXJ2ZUFzcGVjdFJhdGlvPSJ4TWlkWU1pZCBtZWV0Ij4NCiAgICA8IS0tIEdlbmVyYXRvcjogU2tldGNoIDMuMy4xICgxMjAwNSkgLSBodHRwOi8vd3d3LmJvaGVtaWFuY29kaW5nLmNvbS9za2V0Y2ggLS0+DQogICAgPHRpdGxlPlNsaWNlIDE8L3RpdGxlPg0KICAgIDxkZXNjPkNyZWF0ZWQgd2l0aCBTa2V0Y2guPC9kZXNjPg0KICAgIDxkZWZzPjwvZGVmcz4NCiAgICA8ZyBpZD0iUGFnZS0xIiBzdHJva2U9Im5vbmUiIHN0cm9rZS13aWR0aD0iMSIgZmlsbD0ibm9uZSIgZmlsbC1ydWxlPSJldmVub2RkIiBza2V0Y2g6dHlwZT0iTVNQYWdlIj4NCiAgICAgICAgPGcgaWQ9InZpc2EiIHNrZXRjaDp0eXBlPSJNU0xheWVyR3JvdXAiPg0KICAgICAgICAgICAgPHJlY3QgaWQ9IlJlY3RhbmdsZS0xIiBmaWxsPSIjMEU0NTk1IiBza2V0Y2g6dHlwZT0iTVNTaGFwZUdyb3VwIiB4PSIwIiB5PSIwIiB3aWR0aD0iNzUwIiBoZWlnaHQ9IjQ3MSIgcng9IjQwIj48L3JlY3Q+DQogICAgICAgICAgICA8cGF0aCBkPSJNMjc4LjE5NzUsMzM0LjIyNzUgTDMxMS41NTg1LDEzOC40NjU1IEwzNjQuOTE3NSwxMzguNDY1NSBMMzMxLjUzMzUsMzM0LjIyNzUgTDI3OC4xOTc1LDMzNC4yMjc1IEwyNzguMTk3NSwzMzQuMjI3NSBaIiBpZD0iU2hhcGUiIGZpbGw9IiNGRkZGRkYiIHNrZXRjaDp0eXBlPSJNU1NoYXBlR3JvdXAiPjwvcGF0aD4NCiAgICAgICAgICAgIDxwYXRoIGQ9Ik01MjQuMzA3NSwxNDIuNjg3NSBDNTEzLjczNTUsMTM4LjcyMTUgNDk3LjE3MTUsMTM0LjQ2NTUgNDc2LjQ4NDUsMTM0LjQ2NTUgQzQyMy43NjA1LDEzNC40NjU1IDM4Ni42MjA1LDE2MS4wMTY1IDM4Ni4zMDQ1LDE5OS4wNjk1IEMzODYuMDA3NSwyMjcuMTk4NSA0MTIuODE4NSwyNDIuODkwNSA0MzMuMDU4NSwyNTIuMjU0NSBDNDUzLjgyNzUsMjYxLjg0OTUgNDYwLjgxMDUsMjY3Ljk2OTUgNDYwLjcxMTUsMjc2LjUzNzUgQzQ2MC41Nzk1LDI4OS42NTk1IDQ0NC4xMjU1LDI5NS42NTQ1IDQyOC43ODg1LDI5NS42NTQ1IEM0MDcuNDMxNSwyOTUuNjU0NSAzOTYuMDg1NSwyOTIuNjg3NSAzNzguNTYyNSwyODUuMzc4NSBMMzcxLjY4NjUsMjgyLjI2NjUgTDM2NC4xOTc1LDMyNi4wOTA1IEMzNzYuNjYwNSwzMzEuNTU0NSAzOTkuNzA2NSwzMzYuMjg5NSA0MjMuNjM1NSwzMzYuNTM0NSBDNDc5LjcyNDUsMzM2LjUzNDUgNTE2LjEzNjUsMzEwLjI4NzUgNTE2LjU1MDUsMjY5LjY1MjUgQzUxNi43NTE1LDI0Ny4zODM1IDUwMi41MzU1LDIzMC40MzU1IDQ3MS43NTE1LDIxNi40NjQ1IEM0NTMuMTAwNSwyMDcuNDA4NSA0NDEuNjc4NSwyMDEuMzY1NSA0NDEuNzk5NSwxOTIuMTk1NSBDNDQxLjc5OTUsMTg0LjA1ODUgNDUxLjQ2NzUsMTc1LjM1NzUgNDcyLjM1NjUsMTc1LjM1NzUgQzQ4OS44MDU1LDE3NS4wODY1IDUwMi40NDQ1LDE3OC44OTE1IDUxMi4yOTI1LDE4Mi44NTc1IEw1MTcuMDc0NSwxODUuMTE2NSBMNTI0LjMwNzUsMTQyLjY4NzUiIGlkPSJwYXRoMTMiIGZpbGw9IiNGRkZGRkYiIHNrZXRjaDp0eXBlPSJNU1NoYXBlR3JvdXAiPjwvcGF0aD4NCiAgICAgICAgICAgIDxwYXRoIGQ9Ik02NjEuNjE0NSwxMzguNDY1NSBMNjIwLjM4MzUsMTM4LjQ2NTUgQzYwNy42MTA1LDEzOC40NjU1IDU5OC4wNTI1LDE0MS45NTE1IDU5Mi40NDI1LDE1NC42OTk1IEw1MTMuMTk3NSwzMzQuMTAyNSBMNTY5LjIyODUsMzM0LjEwMjUgQzU2OS4yMjg1LDMzNC4xMDI1IDU3OC4zOTA1LDMwOS45ODA1IDU4MC40NjI1LDMwNC42ODQ1IEM1ODYuNTg1NSwzMDQuNjg0NSA2NDEuMDE2NSwzMDQuNzY4NSA2NDguNzk4NSwzMDQuNzY4NSBDNjUwLjM5NDUsMzExLjYyMTUgNjU1LjI5MDUsMzM0LjEwMjUgNjU1LjI5MDUsMzM0LjEwMjUgTDcwNC44MDI1LDMzNC4xMDI1IEw2NjEuNjE0NSwxMzguNDY1NSBMNjYxLjYxNDUsMTM4LjQ2NTUgWiBNNTk2LjE5NzUsMjY0Ljg3MjUgQzYwMC42MTA1LDI1My41OTM1IDYxNy40NTY1LDIxMC4xNDk1IDYxNy40NTY1LDIxMC4xNDk1IEM2MTcuMTQxNSwyMTAuNjcwNSA2MjEuODM2NSwxOTguODE1NSA2MjQuNTMxNSwxOTEuNDY1NSBMNjI4LjEzODUsMjA4LjM0MzUgQzYyOC4xMzg1LDIwOC4zNDM1IDYzOC4zNTU1LDI1NS4wNzI1IDY0MC40OTA1LDI2NC44NzE1IEw1OTYuMTk3NSwyNjQuODcxNSBMNTk2LjE5NzUsMjY0Ljg3MjUgTDU5Ni4xOTc1LDI2NC44NzI1IFoiIGlkPSJQYXRoIiBmaWxsPSIjRkZGRkZGIiBza2V0Y2g6dHlwZT0iTVNTaGFwZUdyb3VwIj48L3BhdGg+DQogICAgICAgICAgICA8cGF0aCBkPSJNMjMyLjkwMjUsMTM4LjQ2NTUgTDE4MC42NjI1LDI3MS45NjA1IEwxNzUuMDk2NSwyNDQuODMxNSBDMTY1LjM3MTUsMjEzLjU1NzUgMTM1LjA3MTUsMTc5LjY3NTUgMTAxLjE5NzUsMTYyLjcxMjUgTDE0OC45NjQ1LDMzMy45MTU1IEwyMDUuNDE5NSwzMzMuODUwNSBMMjg5LjQyMzUsMTM4LjQ2NTUgTDIzMi45MDI1LDEzOC40NjU1IiBpZD0icGF0aDE2IiBmaWxsPSIjRkZGRkZGIiBza2V0Y2g6dHlwZT0iTVNTaGFwZUdyb3VwIj48L3BhdGg+DQogICAgICAgICAgICA8cGF0aCBkPSJNMTMxLjkxOTUsMTM4LjQ2NTUgTDQ1Ljg3ODUsMTM4LjQ2NTUgTDQ1LjE5NzUsMTQyLjUzODUgQzExMi4xMzY1LDE1OC43NDI1IDE1Ni40Mjk1LDE5Ny45MDE1IDE3NC44MTU1LDI0NC45NTI1IEwxNTYuMTA2NSwxNTQuOTkyNSBDMTUyLjg3NjUsMTQyLjU5NjUgMTQzLjUwODUsMTM4Ljg5NzUgMTMxLjkxOTUsMTM4LjQ2NTUiIGlkPSJwYXRoMTgiIGZpbGw9IiNGMkFFMTQiIHNrZXRjaDp0eXBlPSJNU1NoYXBlR3JvdXAiPjwvcGF0aD4NCiAgICAgICAgPC9nPg0KICAgIDwvZz4NCjwvc3ZnPg=="
-            )
-        ),
-        MASTER_CARD(
-            iconName = "mastercard",
-            base64Data = lazyOf(
-                "PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiIHN0YW5kYWxvbmU9Im5vIj8+DQo8c3ZnIHZpZXdCb3g9IjAgMCA3NTAgNDcxIiB2ZXJzaW9uPSIxLjEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHByZXNlcnZlQXNwZWN0UmF0aW89InhNaWRZTWlkIG1lZXQiPg0KICAgIDwhLS0gR2VuZXJhdG9yOiBTa2V0Y2ggMzkuMSAoMzE3MjApIC0gaHR0cDovL3d3dy5ib2hlbWlhbmNvZGluZy5jb20vc2tldGNoIC0tPg0KICAgIDx0aXRsZT5TbGljZSAxPC90aXRsZT4NCiAgICA8ZGVzYz5DcmVhdGVkIHdpdGggU2tldGNoLjwvZGVzYz4NCiAgICA8ZGVmcz48L2RlZnM+DQogICAgPGcgaWQ9IlBhZ2UtMSIgc3Ryb2tlPSJub25lIiBzdHJva2Utd2lkdGg9IjEiIGZpbGw9Im5vbmUiIGZpbGwtcnVsZT0iZXZlbm9kZCI+DQogICAgICAgIDxnIGlkPSJtYXN0ZXJjYXJkIj4NCiAgICAgICAgICAgIDxyZWN0IGlkPSJSZWN0YW5nbGUtMSIgZmlsbD0iI0Y0RjRGNCIgeD0iMCIgeT0iMCIgd2lkdGg9Ijc1MCIgaGVpZ2h0PSI0NzEiIHJ4PSI0MCI+PC9yZWN0Pg0KICAgICAgICAgICAgPGcgaWQ9Im1hcmsiIHRyYW5zZm9ybT0idHJhbnNsYXRlKDEyNS43MTk5OTcsIDQxLjg1MDg2MikiPg0KICAgICAgICAgICAgICAgIDxnIGlkPSJ0ZXh0IiB0cmFuc2Zvcm09InRyYW5zbGF0ZSgyNS4xNDI2NzksIDMyOC4zNjAwMjIpIiBmaWxsPSIjMDAwMDAwIj4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTQ2Ny43MTU1NjEsNTEuOTMyNjg5OSBDNDY2LjUwMjYwNCw1MS45NjIzNTg1IDQ2NS41MDM0MDUsNTIuMzY0ODk0OCA0NjQuNzE3OTYyLDUzLjE0MDMwMDEgQzQ2My45MzI1MTYsNTMuOTE1NzMyMSA0NjMuNTI2MDI3LDU0Ljg4NjEwOTggNDYzLjQ5ODQ5NCw1Ni4wNTE0MzYyIEM0NjMuNTI2MDI3LDU3LjIwNzk0OTcgNDYzLjkzMjUxNiw1OC4xNzU4MDM2IDQ2NC43MTc5NjMsNTguOTU1MDAwNSBDNDY1LjUwMzQwNiw1OS43MzQyMTI1IDQ2Ni41MDI2MDQsNjAuMTM5MjcyNiA0NjcuNzE1NTYxLDYwLjE3MDE4MjUgQzQ2OC45MDA3NjQsNjAuMTM5MjcyNiA0NjkuODg3MzUyLDU5LjczNDIxMjMgNDcwLjY3NTMyNiw1OC45NTUwMDAyIEM0NzEuNDYzMjg1LDU4LjE3NTgwMyA0NzEuODcyMjk3LDU3LjIwNzk0OTMgNDcxLjkwMjM2Miw1Ni4wNTE0MzYyIEM0NzEuODcyOTI3LDU0Ljg4NjEwOTggNDcxLjQ2NTE3Nyw1My45MTU3MzIgNDcwLjY3OTEwOSw1My4xNDAyOTk4IEM0NjkuODkzMDI2LDUyLjM2NDg5NDMgNDY4LjkwNTE3OCw1MS45NjIzNTgxIDQ2Ny43MTU1NjEsNTEuOTMyNjg5OSBMNDY3LjcxNTU2MSw1MS45MzI2ODk5IEw0NjcuNzE1NTYxLDUxLjkzMjY4OTkgWiBNNDY3LjcxNTU2MSw1OS4yNjE2MzU1IEM0NjYuNzkxMzkyLDU5LjIzODkyOTIgNDY2LjAyOTI3Nyw1OC45MjU5ODU0IDQ2NS40MjkyMTQsNTguMzIyODAzMyBDNDY0LjgyOTE0NSw1Ny43MTk2Mzc0IDQ2NC41MTg0OTksNTYuOTYyNTE1OSA0NjQuNDk3MjczLDU2LjA1MTQzNjIgQzQ2NC41MTg0OTksNTUuMTM2MzgwNCA0NjQuODI5MTQ2LDU0LjM3OTY3OSA0NjUuNDI5MjE0LDUzLjc4MTMzIEM0NjYuMDI5Mjc3LDUzLjE4MzAwNjIgNDY2Ljc5MTM5Miw1Mi44NzMwMDcxIDQ2Ny43MTU1NjEsNTIuODUxMzMxOCBDNDY4LjYyMDM4Myw1Mi44NzMwMDc2IDQ2OS4zNzA3MjgsNTMuMTgzMDA2NiA0NjkuOTY2NTk3LDUzLjc4MTMzMDIgQzQ3MC41NjI0NTIsNTQuMzc5Njc5IDQ3MC44NzE0MTcsNTUuMTM2MzgwNCA0NzAuODkzNDk0LDU2LjA1MTQzNjIgQzQ3MC44NzE0MTcsNTYuOTYyNTE2MSA0NzAuNTYyNDUyLDU3LjcxOTYzNzggNDY5Ljk2NjU5Nyw1OC4zMjI4MDMzIEM0NjkuMzcwNzI4LDU4LjkyNTk4NSA0NjguNjIwMzg0LDU5LjIzODkyODcgNDY3LjcxNTU2MSw1OS4yNjE2MzU1IEw0NjcuNzE1NTYxLDU5LjI2MTYzNTUgTDQ2Ny43MTU1NjEsNTkuMjYxNjM1NSBaIE00NjcuOTU3Njg5LDU0LjEyMzI5NzUgTDQ2Ni4xOTIxNyw1NC4xMjMyOTc1IEw0NjYuMTkyMTcsNTcuOTQ5Mjg5OSBMNDY3LjAwOTM1Myw1Ny45NDkyODk5IEw0NjcuMDA5MzUzLDU2LjUxNTgwNDYgTDQ2Ny4zODI2MzQsNTYuNTE1ODA0NiBMNDY4LjU0MjgzMiw1Ny45NDkyODk5IEw0NjkuNTIxNDM0LDU3Ljk0OTI4OTkgTDQ2OC4yNzA0MzgsNTYuNTA1NzA5NyBDNDY4LjY2MTE1OCw1Ni40NTY3MTY5IDQ2OC45NjE3MTYsNTYuMzMwMTA5IDQ2OS4xNzIxMTMsNTYuMTI1ODg2MSBDNDY5LjM4MjQ5OCw1NS45MjE2ODM2IDQ2OS40ODg4NDksNTUuNjYxMzE4MSA0NjkuNDkxMTY4LDU1LjM0NDc4ODUgQzQ2OS40ODg0MjksNTQuOTY3MDc5NiA0NjkuMzU1MTc0LDU0LjY3MDExOTUgNDY5LjA5MTQwNCw1NC40NTM5MDczIEM0NjguODI3NjIxLDU0LjIzNzcxOSA0NjguNDQ5NzE3LDU0LjEyNzUxNTggNDY3Ljk1NzY4OSw1NC4xMjMyOTc1IEw0NjcuOTU3Njg5LDU0LjEyMzI5NzUgTDQ2Ny45NTc2ODksNTQuMTIzMjk3NSBaIE00NjcuOTQ3Niw1NC44NDAwNDAyIEM0NjguMTY2ODEzLDU0Ljg0MDI2MiA0NjguMzM4NzQxLDU0Ljg4Mjc0NTMgNDY4LjQ2MzM4Myw1NC45Njc0ODg1IEM0NjguNTg4MDE1LDU1LjA1MjI1NTIgNDY4LjY1MTQ4OSw1NS4xNzgwMjE4IDQ2OC42NTM4MDgsNTUuMzQ0Nzg4MiBDNDY4LjY1MTQ4Myw1NS41MTY0MTI5IDQ2OC41ODgwMTUsNTUuNjQ1MTIzNSA0NjguNDYzMzgzLDU1LjczMDkyIEM0NjguMzM4NzQxLDU1LjgxNjczODUgNDY4LjE2NjgxMyw1NS44NTk2NDEyIDQ2Ny45NDc2LDU1Ljg1OTYzMSBMNDY3LjAwOTM1Myw1NS44NTk2MzEgTDQ2Ny4wMDkzNTMsNTQuODQwMDM5MyBMNDY3Ljk0NzYsNTQuODQwMDQwMiBMNDY3Ljk0NzYsNTQuODQwMDQwMiBaIiBpZD0icGF0aDMwNzgiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTkuMzQzMzE3MjQsNTcuNTQyODAyOSBMMC41ODgxNzU5MTYsNTcuNTQyODAyOSBMMC41ODgxNzU5MTYsMTYuNjYwMDA0NSBMOS4xNzE2NDc1NywxNi42NjAwMDQ1IEw5LjE3MTY0NzU3LDIxLjY0MTUxODYgQzkuMTcxNjQ3NTcsMjEuNjQxNTE4NiAxNi43MTA3MzU1LDE1LjU1MzI0ODUgMjEuMTg4NTA4MywxNS42MjkzNTA4IEMyOS44OTQ5Mjk4LDE1Ljc3NzMzMTcgMzUuMDkzNzI5LDIzLjE4NzUwOTggMzUuMDkzNzI5LDIzLjE4NzUwOTggQzM1LjA5MzcyOSwyMy4xODc1MDk4IDM5LjMxMDk4OTMsMTUuNjI5MzUwOCA0OC44MjcyOTE4LDE1LjYyOTM1MDggQzYyLjg5OTc5ODgsMTUuNjI5MzUwOCA2NC45NjQyMTQ5LDI4LjUxMjU4NTggNjQuOTY0MjE0OSwyOC41MTI1ODU4IEw2NC45NjQyMTQ5LDU3LjM3MTAyNzMgTDU2LjU1MjQxMDgsNTcuMzcxMDI3MyBMNTYuNTUyNDEwOCwzMS45NDgxMDg3IEM1Ni41NTI0MTA4LDMxLjk0ODEwODcgNTYuNTgyNTkyMiwyNC4yMTgxNzQxIDQ3LjQ1MzkzMjMsMjQuMjE4MTc0MSBDMzguMDEzOTc0NywyNC4yMTgxNzQxIDM3LjE1Mzc2MjksMzEuOTQ4MTA4NyAzNy4xNTM3NjI5LDMxLjk0ODEwODcgTDM3LjE1Mzc2MjksNTcuMzcxMDI3MyBMMjguMzk4NjIxNSw1Ny4zNzEwMjczIEwyOC4zOTg2MjE1LDMxLjc3NjMzMzEgQzI4LjM5ODYyMTUsMzEuNzc2MzMzMSAyNy41NTc1NDk2LDIzLjcwMjgzNjYgMTkuNjQzNDgzNCwyMy43MDI4MzY2IEM5LjM2NTAxMTMsMjMuNzAyODM2NiA5LjE3MTY0NzU3LDMxLjk0ODEwODcgOS4xNzE2NDc1NywzMS45NDgxMDg3IEw5LjM0MzMxNzI0LDU3LjU0MjgwMjkgTDkuMzQzMzE3MjQsNTcuNTQyODAyOSBaIiBpZD0icGF0aDMwMDYiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTI3NS41OTY4OTgsMTUuNjIzNzczIEMyNzEuMTE5MTIyLDE1LjU0NzY4MTQgMjYzLjU4MDc0MSwyMS42MzU1OTY3IDI2My41ODA3NDEsMjEuNjM1NTk2NyBMMjYzLjU4MDc0MSwxNi42NjQ5MjY4IEwyNTQuOTg4MTgyLDE2LjY2NDkyNjggTDI1NC45ODgxODIsNTcuNTM4NjExNSBMMjYzLjc0ODU2NSw1Ny41Mzg2MTE1IEwyNjMuNTgwNzQxLDMxLjk0NjM3ODMgQzI2My41ODA3NDEsMzEuOTQ2Mzc4MyAyNjMuNzc0NDUsMjMuNzE3OTA0NCAyNzQuMDUyOTIzLDIzLjcxNzkwNDQgQzI3NS45NjE4MjQsMjMuNzE3OTA0NCAyNzcuNDQ0MzYzLDI0LjE4MDU2OSAyNzguNjE3NzIsMjQuODkzNDAwNyBMMjc4LjYxNzcyLDI0Ljg1OTg2ODggTDI4MS40NzA3MTgsMTYuOTAwMDc5OCBDMjc5Ljc0OTA5MiwxNi4xNzUwMTc2IDI3Ny43OTEyMjMsMTUuNjYxMDY2NCAyNzUuNTk2ODk4LDE1LjYyMzc3MyBMMjc1LjU5Njg5OCwxNS42MjM3NzMgTDI3NS41OTY4OTgsMTUuNjIzNzczIFoiIGlkPSJwYXRoMzAwOCI+PC9wYXRoPg0KICAgICAgICAgICAgICAgICAgICA8cGF0aCBkPSJNMzk4LjkyNzc0LDE1LjYyMzc3MyBDMzk0LjQ0OTk2NCwxNS41NDc2ODE0IDM4Ni45MTE1ODIsMjEuNjM1NTk2NyAzODYuOTExNTgyLDIxLjYzNTU5NjcgTDM4Ni45MTE1ODIsMTYuNjY0OTI2OCBMMzc4LjMxOTAyMywxNi42NjQ5MjY4IEwzNzguMzE5MDIzLDU3LjUzODYxMTUgTDM4Ny4wNzk0MDYsNTcuNTM4NjExNSBMMzg2LjkxMTU4MiwzMS45NDYzNzgzIEMzODYuOTExNTgyLDMxLjk0NjM3ODMgMzg3LjEwNTI5MSwyMy43MTc5MDQ0IDM5Ny4zODM3NjQsMjMuNzE3OTA0NCBDMzk5LjI5MjY2NiwyMy43MTc5MDQ0IDQwMC43NzUyMDQsMjQuMTgwNTY5IDQwMS45NDg1NjEsMjQuODkzNDAwNyBMNDAxLjk0ODU2MSwyNC44NTk4Njg4IEw0MDQuODAxNTU5LDE2LjkwMDA3OTggQzQwMy4wNzk5MzMsMTYuMTc1MDcxMyA0MDEuMTIyMDY0LDE1LjY2MTEyMDEgMzk4LjkyNzc0LDE1LjYyMzgyNjcgTDM5OC45Mjc3NCwxNS42MjM3NzMgTDM5OC45Mjc3NCwxNS42MjM3NzMgWiIgaWQ9InBhdGgzMDEzIj48L3BhdGg+DQogICAgICAgICAgICAgICAgICAgIDxwYXRoIGQ9Ik05My4yNzM1Mjk1LDE1LjQ1NTg0NDkgQzgwLjE3MDg2NDYsMTUuNDU1ODQ0OSA3My4yMzY4NjI2LDI3LjIzOTY4NTkgNzMuMjAxODQ3OSwzNy4wODQ5NzYzIEM3My4xNjU4NjY2LDQ3LjE3NjI3NDYgODEuMDk1NTk1OSw1OC44MTQ4NjQ2IDkzLjY0Mjc0MTEsNTguODE0ODY0NiBDMTAwLjk2MjY3OCw1OC44MTQ4NjQ2IDEwNi45NzYwNDEsNTMuNDA3NTgxNyAxMDYuOTc2MDQxLDUzLjQwNzU4MTcgTDEwNi45NjAxNDUsNTcuNTcyMTk3MSBMMTE1LjU3Nzk5OCw1Ny41NzIxOTcxIEwxMTUuNTc3OTk4LDE2LjY0ODgyNzIgTDEwNi45Mjk3OTIsMTYuNjQ4ODI3MiBMMTA2LjkyOTc5MiwyMS44MDM1MjQ4IEMxMDYuOTI5NzkyLDIxLjgwMzUyNDggMTAxLjI4MjY1NCwxNS40NTU4NDQ5IDkzLjI3MzU3MjUsMTUuNDU1ODQ0OSBMOTMuMjczNTI5NSwxNS40NTU4NDQ5IEw5My4yNzM1Mjk1LDE1LjQ1NTg0NDkgWiBNOTQuOTUxNzYzOCwyMy43ODUwNzU2IEMxMDEuOTkxNDMzLDIzLjc4NTA3NTYgMTA3LjcwNjM0NCwyOS45MTIyOTQyIDEwNy43MDYzNDQsMzcuNDU0NDE4IEMxMDcuNzA2MzQ0LDQ0Ljk5NjU0MTggMTAxLjk5MTQzMyw1MS4wOTAxNzQ4IDk0Ljk1MTc2MzgsNTEuMDkwMTc0OCBDODcuOTEyMDk0Nyw1MS4wOTAxNzQ4IDgyLjIzMDc0ODIsNDQuOTk2NTQxOCA4Mi4yMzA3NDgyLDM3LjQ1NDQxOCBDODIuMjMwNzQ4MiwyOS45MTIyOTQyIDg3LjkxMjA5NDcsMjMuNzg1MDc1NiA5NC45NTE3NjM4LDIzLjc4NTA3NTYgTDk0Ljk1MTc2MzgsMjMuNzg1MDc1NiBMOTQuOTUxNzYzOCwyMy43ODUwNzU2IFoiIGlkPSJwYXRoMzAxNSI+PC9wYXRoPg0KICAgICAgICAgICAgICAgICAgICA8cGF0aCBkPSJNMzQ0LjU5NzU3OCwxNS40NTU4NDQ5IEMzMzEuNDk0OTEzLDE1LjQ1NTg0NDkgMzI0LjU2MDkxMSwyNy4yMzk2ODU5IDMyNC41MjU4OTYsMzcuMDg0OTc2MyBDMzI0LjQ4OTkxNSw0Ny4xNzYyNzQ2IDMzMi40MTk2NDQsNTguODE0ODY0NiAzNDQuOTY2Nzg5LDU4LjgxNDg2NDYgQzM1Mi4yODY3MjYsNTguODE0ODY0NiAzNTguMzAwMDg5LDUzLjQwNzU4MTcgMzU4LjMwMDA4OSw1My40MDc1ODE3IEwzNTguMjg0MTkzLDU3LjU3MjE5NzEgTDM2Ni45MDIwNDYsNTcuNTcyMTk3MSBMMzY2LjkwMjA0NiwxNi42NDg4MjcyIEwzNTguMjUzODQsMTYuNjQ4ODI3MiBMMzU4LjI1Mzg0LDIxLjgwMzUyNDggQzM1OC4yNTM4NCwyMS44MDM1MjQ4IDM1Mi42MDY3MDIsMTUuNDU1ODQ0OSAzNDQuNTk3NjIsMTUuNDU1ODQ0OSBMMzQ0LjU5NzU3OCwxNS40NTU4NDQ5IEwzNDQuNTk3NTc4LDE1LjQ1NTg0NDkgWiBNMzQ2LjI3NTgxMiwyMy43ODUwNzU2IEMzNTMuMzE1NDgxLDIzLjc4NTA3NTYgMzU5LjAzMDM5MiwyOS45MTIyOTQyIDM1OS4wMzAzOTIsMzcuNDU0NDE4IEMzNTkuMDMwMzkyLDQ0Ljk5NjU0MTggMzUzLjMxNTQ4MSw1MS4wOTAxNzQ4IDM0Ni4yNzU4MTIsNTEuMDkwMTc0OCBDMzM5LjIzNjE0Myw1MS4wOTAxNzQ4IDMzMy41NTQ3OTYsNDQuOTk2NTQxOCAzMzMuNTU0Nzk2LDM3LjQ1NDQxOCBDMzMzLjU1NDc5NiwyOS45MTIyOTQyIDMzOS4yMzYxNDMsMjMuNzg1MDc1NiAzNDYuMjc1ODEyLDIzLjc4NTA3NTYgTDM0Ni4yNzU4MTIsMjMuNzg1MDc1NiBMMzQ2LjI3NTgxMiwyMy43ODUwNzU2IFoiIGlkPSJwYXRoMzAyMCI+PC9wYXRoPg0KICAgICAgICAgICAgICAgICAgICA8cGF0aCBkPSJNNDI3LjM0MjI0OSwxNS40NTU4NDQ5IEM0MTQuMjM5NTg0LDE1LjQ1NTg0NDkgNDA3LjMwNTU4MiwyNy4yMzk2ODU5IDQwNy4yNzA1NjcsMzcuMDg0OTc2MyBDNDA3LjIzNDU4Niw0Ny4xNzYyNzQ2IDQxNS4xNjQzMTUsNTguODE0ODY0NiA0MjcuNzExNDYsNTguODE0ODY0NiBDNDM1LjAzMTM5Nyw1OC44MTQ4NjQ2IDQ0MS4wNDQ3Niw1My40MDc1ODE3IDQ0MS4wNDQ3Niw1My40MDc1ODE3IEw0NDEuMDI4ODY0LDU3LjU3MjE5NzEgTDQ0OS42NDY3MTgsNTcuNTcyMTk3MSBMNDQ5LjY0NjcxOCwwLjQ5NDA3NDYyIEw0NDAuOTk4NTExLDAuNDk0MDc0NjIgTDQ0MC45OTg1MTEsMjEuODAzNTI0OCBDNDQwLjk5ODUxMSwyMS44MDM1MjQ4IDQzNS4zNTEzNzMsMTUuNDU1ODQ0OSA0MjcuMzQyMjkyLDE1LjQ1NTg0NDkgTDQyNy4zNDIyNDksMTUuNDU1ODQ0OSBMNDI3LjM0MjI0OSwxNS40NTU4NDQ5IFogTTQyOS4wMjA0ODMsMjMuNzg1MDc1NiBDNDM2LjA2MDE1MiwyMy43ODUwNzU2IDQ0MS43NzUwNjMsMjkuOTEyMjk0MiA0NDEuNzc1MDYzLDM3LjQ1NDQxOCBDNDQxLjc3NTA2Myw0NC45OTY1NDE4IDQzNi4wNjAxNTIsNTEuMDkwMTc0OCA0MjkuMDIwNDgzLDUxLjA5MDE3NDggQzQyMS45ODA4MTQsNTEuMDkwMTc0OCA0MTYuMjk5NDY3LDQ0Ljk5NjU0MTggNDE2LjI5OTQ2NywzNy40NTQ0MTggQzQxNi4yOTk0NjcsMjkuOTEyMjk0MiA0MjEuOTgwODE0LDIzLjc4NTA3NTYgNDI5LjAyMDQ4MywyMy43ODUwNzU2IEw0MjkuMDIwNDgzLDIzLjc4NTA3NTYgTDQyOS4wMjA0ODMsMjMuNzg1MDc1NiBaIiBpZD0icGF0aDMwMjIiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTE0MS44NzIxMjIsNTguOTE3MDA3OCBDMTMyLjk0NTU4LDU4LjkxNzAwNzggMTI0LjcwNTE3Niw1My40MjAxNjY5IDEyNC43MDUxNzYsNTMuNDIwMTY2OSBMMTI4LjQ4MTkwNyw0Ny41Nzk3NjQxIEMxMjguNDgxOTA3LDQ3LjU3OTc2NDEgMTM2LjI3ODk3OCw1MS4xODcwNzMzIDE0MS44NzIxMjIsNTEuMTg3MDczMyBDMTQ1LjUwNjEzLDUxLjE4NzA3MzMgMTUxLjU4MzkzNyw1MC4wMTI4NjY3IDE1MS42NTcyNzQsNDYuMzc3MzM0OCBDMTUxLjczNDgyMiw0Mi41MzQ5NDc4IDE0MS40NDI5NDUsNDEuMzk1ODEgMTQxLjQ0Mjk0NSw0MS4zOTU4MSBDMTQxLjQ0Mjk0NSw0MS4zOTU4MSAxMjYuMDc4NTM2LDQxLjE4NjA4NTMgMTI2LjA3ODUzNiwyOC41MTI1ODU4IEMxMjYuMDc4NTM2LDIwLjU0MjEyNDYgMTMzLjc1MTkzOCwxNS40NTc1NzUyIDE0My41ODg4MTgsMTUuNDU3NTc1MiBDMTQ5LjI3MjY2NywxNS40NTc1NzUyIDE1OS44OTc0MSwyMC40MzkwODkzIDE1OS44OTc0MSwyMC40MzkwODkzIEwxNTUuNjA1Njc0LDI3LjEzODM3MDIgQzE1NS42MDU2NzQsMjcuMTM4MzcwMiAxNDcuNDAyMjE4LDIzLjg1ODkyMSAxNDMuMDczODAyLDIzLjcwMjgzNjYgQzEzOS40MTg4MDYsMjMuNTcxMDQxMyAxMzUuMDA1MzQ2LDI1LjMyMjE0NjUgMTM1LjAwNTM0NiwyOC41MTI1ODU4IEMxMzUuMDA1MzQ2LDM3LjE4MDY5MjYgMTYwLjU4NDA4NCwyNy44MzcxOTggMTYwLjU4NDA4NCw0NS4zNDY2NzA0IEMxNjAuNTg0MDg0LDU2LjgzMzgxODggMTUwLjE2NjY5MSw1OC45MTcwMDc4IDE0MS44NzIxMjIsNTguOTE3MDA3OCBMMTQxLjg3MjEyMiw1OC45MTcwMDc4IEwxNDEuODcyMTIyLDU4LjkxNzAwNzggWiIgaWQ9InBhdGgzMDI0Ij48L3BhdGg+DQogICAgICAgICAgICAgICAgICAgIDxwYXRoIGQ9Ik0xNzQuODAyMTQ5LDQuODA5MjA3MjQgTDE3NC44MDIxNDksMTYuNjk4NTEyNCBMMTY3LjE4Mjk2NiwxNi42OTg1MTI0IEwxNjcuMTgyOTY2LDI1LjI5NjQyOCBMMTc0LjgwMjE0OSwyNS4yOTY0MjggTDE3NC44MDIxNDksNDUuODUwODIgQzE3NC44MDIxNDksNDUuODUwODIgMTc0LjEyNzgyNyw1OS43NTUyNjE2IDE4OS4wNjcxNDEsNTkuNzU1MjYxNiBDMTkzLjE5NzUzLDU5Ljc1NTI2MTYgMjAxLjI4NDY4Niw1Ni42OTg5NzEzIDIwMS4yODQ2ODYsNTYuNjk4OTcxMyBMMTk3LjgyNzUyMyw0Ny43NjUxOTk2IEMxOTcuODI3NTIzLDQ3Ljc2NTE5OTYgMTk0LjYxMTQ2OCw1MC41MTAyNDU0IDE5MC45ODAzMjgsNTAuNDE4NDYyNiBDMTg0LjA3NjE5NSw1MC4yNDQwMjE3IDE4NC4yNjczOTEsNDUuODE3MjM0MyAxODQuMjY3MzkxLDQ1LjgxNzIzNDMgTDE4NC4yNjczOTEsMjUuMjk2NDI4IEwxOTguNDk4ODE3LDI1LjI5NjQyOCBMMTk4LjQ5ODgxNywxNi42OTg1MTI0IEwxODQuMjY3MzkxLDE2LjY5ODUxMjQgTDE4NC4yNjczOTEsNC44MDkyMDcyNCBMMTc0LjgwMjE0OSw0LjgwOTIwNzI0IEwxNzQuODAyMTQ5LDQuODA5MjA3MjQgTDE3NC44MDIxNDksNC44MDkyMDcyNCBaIiBpZD0icGF0aDMwMjYiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTIyNi42NTk1ODgsMTUuOTU5NjI5IEMyMTIuNjEwMDg3LDE1Ljk1OTYyOSAyMDUuNTkwNDE3LDI3LjUzODk3OTMgMjA1LjY0ODA5NSwzNy41ODg3NjA0IEMyMDUuNzA3Mzg0LDQ3LjkyMzg0MTkgMjEyLjA0MDMwNCw1OS41NTM3NDc5IDIyNy40OTg3MDUsNTkuNTUzNzQ3OSBDMjM0LjExNTA3Miw1OS41NTM3NDc5IDI0My40MDgzNjYsNTMuNzQzNDM3OCAyNDMuNDA4MzY2LDUzLjc0MzQzNzggTDIzOS40MTQxNjgsNDYuNzkxMjE3IEMyMzkuNDE0MTY4LDQ2Ljc5MTIxNyAyMzMuMDcyNTQ4LDUxLjI5MTY4ODQgMjI3LjQ5ODcwNSw1MS4yOTE2ODg0IEMyMTYuMzM5MTcyLDUxLjI5MTY4ODQgMjE1LjYxNjgwNiw0MC4zNzYzNjU5IDIxNS42MTY4MDYsNDAuMzc2MzY1OSBMMjQ1LjQ4OTM3Niw0MC4zNzYzNjU5IEMyNDUuNDg5Mzc2LDQwLjM3NjM2NTkgMjQ3LjcxNzk4NSwxNS45NTk2MjkgMjI2LjY1OTU4OCwxNS45NTk2MjkgTDIyNi42NTk1ODgsMTUuOTU5NjI5IEwyMjYuNjU5NTg4LDE1Ljk1OTYyOSBaIE0yMjUuMzg0MTMsMjMuOTg2NTg5MyBDMjI1LjcxNTQxNiwyMy45Njc3ODEzIDIyNi4wNzA1NjgsMjMuOTg2NTg5MyAyMjYuNDI0NjM1LDIzLjk4NjU4OTMgQzIzNi45Mzc5NTQsMjMuOTg2NTg5MyAyMzYuODYzMjUyLDMzLjkyNzkyOTIgMjM2Ljg2MzI1MiwzMy45Mjc5MjkyIEwyMTUuNjE2ODA2LDMzLjkyNzkyOTIgQzIxNS42MTY4MDYsMzMuOTI3OTI5MiAyMTUuMTE0MjA2LDI0LjU3MDc5NjIgMjI1LjM4NDEzLDIzLjk4NjU4OTMgTDIyNS4zODQxMywyMy45ODY1ODkzIEwyMjUuMzg0MTMsMjMuOTg2NTg5MyBaIiBpZD0icGF0aDMwMzQiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICAgICAgPHBhdGggZD0iTTMxNS41MTYyLDQ2LjY4NiBMMzE5LjUyMjAzLDU0LjcwMjY0MjcgQzMxOS41MjIwMyw1NC43MDI2NDI3IDMxMy4xNzMwMiw1OC44MzI0MjU4IDMwNi4wNDc4OTgsNTguODMyNDI1OCBDMjkxLjI5NjU1Nyw1OC44MzI0MjU4IDI4My4xMDU0NDIsNDcuNzIzNDk5NyAyODMuMTA1NDQyLDM3LjIxMTc4NDggQzI4My4xMDU0NDIsMjAuNjkxMjMzOSAyOTYuMTQwNjk4LDE1LjgzNDA2NzIgMzA0Ljk1NTM5NywxNS44MzQwNjcyIEMzMTIuOTU2MzY5LDE1LjgzNDA2NzIgMzE5Ljg4NjE5MywyMC40NDk3MDc3IDMxOS44ODYxOTMsMjAuNDQ5NzA3NyBMMzE1LjM5NDgxOSwyOC40NjYzNTA1IEMzMTUuMzk0ODE5LDI4LjQ2NjM1MDUgMzEyLjY3MTg1OSwyNC4yMTUxMDA0IDMwNC43MTI2MTQsMjQuMjE1MTAwNCBDMjk2Ljc2NjY1NSwyNC4yMTUxMDA0IDI5Mi41NzM3NTUsMzEuMDcwMjE4NSAyOTIuNTczNzU1LDM3LjU3NjE3NTIgQzI5Mi41NzM3NTUsNDQuODY2OTkyNyAyOTcuNDU0MTE0LDUxLjA1ODcwNjQgMzA0LjgzNDAwNSw1MS4wNTg3MDY0IEMzMTAuNjIzMzkzLDUxLjA1ODcwNjQgMzE1LjUxNjIsNDYuNjg2IDMxNS41MTYyLDQ2LjY4NiBMMzE1LjUxNjIsNDYuNjg2IEwzMTUuNTE2Miw0Ni42ODYgWiIgaWQ9InBhdGgzMDM3Ij48L3BhdGg+DQogICAgICAgICAgICAgICAgPC9nPg0KICAgICAgICAgICAgICAgIDxwYXRoIGQ9Ik00OTguNzg3OTg1LDIzNi43ODEyNzkgTDQ5OC43ODc5ODUsMjMxLjI2MDYyMyBMNDk3LjM0NzQ4NCwyMzEuMjYwNjIzIEw0OTUuNjkwNDM2LDIzNS4wNTcyNTIgTDQ5NC4wMzMzODgsMjMxLjI2MDYyMyBMNDkyLjU5Mjg4NiwyMzEuMjYwNjIzIEw0OTIuNTkyODg2LDIzNi43ODEyNzkgTDQ5My42MDk3MTEsMjM2Ljc4MTI3OSBMNDkzLjYwOTcxMSwyMzIuNjE3MjM1IEw0OTUuMTYzMTkzLDIzNi4yMDY2MDMgTDQ5Ni4yMTc2NzgsMjM2LjIwNjYwMyBMNDk3Ljc3MTE2MSwyMzIuNjA3ODE0IEw0OTcuNzcxMTYxLDIzNi43ODEyNzkgTDQ5OC43ODc5ODUsMjM2Ljc4MTI3OSBMNDk4Ljc4Nzk4NSwyMzYuNzgxMjc5IFogTTQ4OS42NjQ4MDcsMjM2Ljc4MTI3OSBMNDg5LjY2NDgwNywyMzIuMjAyNzE1IEw0OTEuNTEwMTU2LDIzMi4yMDI3MTUgTDQ5MS41MTAxNTYsMjMxLjI3MDA0NCBMNDg2LjgxMjA0OSwyMzEuMjcwMDQ0IEw0ODYuODEyMDQ5LDIzMi4yMDI3MTUgTDQ4OC42NTczOTcsMjMyLjIwMjcxNSBMNDg4LjY1NzM5NywyMzYuNzgxMjc5IEw0ODkuNjY0ODA3LDIzNi43ODEyNzkgTDQ4OS42NjQ4MDcsMjM2Ljc4MTI3OSBaIiBpZD0icGF0aDMwNTciIGZpbGw9IiNGNzlGMUEiPjwvcGF0aD4NCiAgICAgICAgICAgICAgICA8cGF0aCBkPSJNNDk5LjA3NjY3OCwxNTQuNzA5ODAyIEM0OTkuMDc2Njc4LDI0MC4xMzUxNTkgNDI5Ljk5OTcwNywzMDkuMzg2MTA1IDM0NC43ODg5MjksMzA5LjM4NjEwNSBDMjU5LjU3ODE1MSwzMDkuMzg2MTA1IDE5MC41MDExNTksMjQwLjEzNTE1OSAxOTAuNTAxMTU5LDE1NC43MDk4MDIgQzE5MC41MDExNTksNjkuMjg0NDMyNiAyNTkuNTc4MTUxLDAuMDMzNDkyMDE3NCAzNDQuNzg4OTI5LDAuMDMzNDkyMDE3NCBDNDI5Ljk5OTcwNywwLjAzMzQ5MjAxNzQgNDk5LjA3NjY3OCw2OS4yODQ0MzI2IDQ5OS4wNzY2NzgsMTU0LjcwOTgwMiBMNDk5LjA3NjY3OCwxNTQuNzA5ODAyIEw0OTkuMDc2Njc4LDE1NC43MDk4MDIgWiIgaWQ9InBhdGgyOTk3IiBmaWxsPSIjRjc5RjFBIj48L3BhdGg+DQogICAgICAgICAgICAgICAgPHBhdGggZD0iTTMwOC43MzkzMiwxNTQuNzA5ODAyIEMzMDguNzM5MzIsMjQwLjEzNTE1OSAyMzkuNjYyMzQ5LDMwOS4zODYxMDUgMTU0LjQ1MTU3MSwzMDkuMzg2MTA1IEM2OS4yNDA3OTMxLDMwOS4zODYxMDUgMC4xNjM4MDEyNzUsMjQwLjEzNTE1OSAwLjE2MzgwMTI3NSwxNTQuNzA5ODAyIEMwLjE2MzgwMTI3NSw2OS4yODQ0MzI2IDY5LjI0MDc5MzEsMC4wMzM0OTIwMTc0IDE1NC40NTE1NzEsMC4wMzM0OTIwMTc0IEMyMzkuNjYyMzQ5LDAuMDMzNDkyMDE3NCAzMDguNzM5MzIsNjkuMjg0NDMyNiAzMDguNzM5MzIsMTU0LjcwOTgwMiBMMzA4LjczOTMyLDE1NC43MDk4MDIgTDMwOC43MzkzMiwxNTQuNzA5ODAyIFoiIGlkPSJwYXRoMjk5NSIgZmlsbD0iI0VBMDAxQiI+PC9wYXRoPg0KICAgICAgICAgICAgICAgIDxwYXRoIGQ9Ik0yNDkuNjIwNTYyLDMyLjk0NzQ4MTIgQzIxMy42MjEzMjYsNjEuMjYzNjgyMyAxOTAuNTEzMTUyLDEwNS4yNjUzNDUgMTkwLjUxMzE1MiwxNTQuNjk1MzA5IEMxOTAuNTEzMTUyLDIwNC4xMjUyNzQgMjEzLjYyMTMyNiwyNDguMTYwNTIgMjQ5LjYyMDU2MiwyNzYuNDc2NzIzIEMyODUuNjE5Nzk5LDI0OC4xNjA1MiAzMDguNzI3OTczLDIwNC4xMjUyNzQgMzA4LjcyNzk3MywxNTQuNjk1MzA5IEMzMDguNzI3OTczLDEwNS4yNjUzNDUgMjg1LjYxOTc5OSw2MS4yNjM2ODIzIDI0OS42MjA1NjIsMzIuOTQ3NDgxMiBMMjQ5LjYyMDU2MiwzMi45NDc0ODEyIEwyNDkuNjIwNTYyLDMyLjk0NzQ4MTIgWiIgaWQ9InBhdGgyOTk5IiBmaWxsPSIjRkY1RjAxIj48L3BhdGg+DQogICAgICAgICAgICA8L2c+DQogICAgICAgIDwvZz4NCiAgICA8L2c+DQo8L3N2Zz4NCg=="
-            )
-        )
-        // TODO cardreader Add support for other card types (see iOS implementation)
-        /* ktlint-enable max-line-length */
-    }
-
-    data class ReceiptData(
-        val purchasedProducts: List<ReceiptLineItem>,
-        val amount: Int,
-        val currency: String,
-        val receiptDate: String,
-        val storeName: String?,
-        val applicationPreferredName: String?,
-        val dedicatedFileName: String?, //  Also known as “AID”.
-        val cardInfo: CardInfo
-    )
-
-    data class CardInfo(
-        val last4CardDigits: String,
-        val brand: CardBrand
-    )
-
-    data class ReceiptLineItem(
-        val title: String,
-        val quantity: String,
-        val amount: String
-    )
 }
